refactor(#237): seperate Column.tsx
