+++ apps/bff/src/erd/vendor.controller.ts
@@ -1,19 +1,20 @@
-import { Controller, Get, Headers, Param } from '@nestjs/common';
+import { Controller, Get, Param } from '@nestjs/common';
 import { VendorService } from './vendor.service';
+import { AuthHeader } from '../common/decorators/auth-header.decorator';
 
 @Controller('api/v1.0')
 export class VendorController {
   constructor(private readonly vendorService: VendorService) {}
 
   @Get('vendors')
-  async listVendors(@Headers('authorization') authHeader: string) {
+  async listVendors(@AuthHeader() authHeader: string) {
     return this.vendorService.listVendors(authHeader);
   }
 
   @Get('vendors/:displayName')
   async getVendor(
     @Param('displayName') displayName: string,
-    @Headers('authorization') authHeader: string,
+    @AuthHeader() authHeader: string,
   ) {
     return this.vendorService.getVendor(displayName, authHeader);
   }
